refactor(inject): split inject.rs into per-platform modules
Move the inline macos/linux/windows modules to inject/{macos,linux,windows}.rs,
matching openlogi-hook's per-platform file layout. Each execute_<os> free
function becomes <os>::execute inside its platform module, the macOS VK_*
constants move next to their only user, and the Linux key-mapping tests move
into linux.rs alongside the code they cover.

The execute/post_horizontal_scroll dispatch now uses cfg_select! so an
unhandled platform is a compile error instead of a silent fall-through.
Helpers whose only callers moved into their module drop pub(super).
